requirements.txt

# Django和REST框架
Django==4.0.2
djangorestframework==3.14.0
django-cors-headers==4.0.0

# 异步HTTP客户端
httpx==0.25.0

# JWT处理
pyjwt==2.8.0

# 环境变量管理
python-dotenv==1.0.0

# ASGI服务器
uvicorn==0.23.2
gunicorn==21.2.0

# 其他工具
python-dateutil==2.8.2
